+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Romantic Versioning](http://dafoster.net/articles/2
 ## [Unreleased]
 ### Added
 - New feature
+- Added zsh autosuggestions plugin (zplug)
+- Added zsh syntax-highlighting plugin (zplug)
+- Added zsh zaw plugin
+- Added git-extra-commands plugin
+- Added git-cal plugin
 
 ### Changed
 - Changes in existing functionality
@@ -18,6 +23,7 @@ and this project adheres to [Romantic Versioning](http://dafoster.net/articles/2
 
 ### Removed
 - Removed feature
+- Removed zsh-syntax-highlighting and zsh-autosuggestions plugins (submodules)
 
 ### Fixed
 - Bugfix